Fixing private method definition, used === instead of ==.
When implementing private methods, the last thing you want to do in a
constructor is to return an object other than "this". Depending on the
environment this could lead to leaks and undefined behavior. Instead, all
priviledged methods (which can access private info) should be defined on "this".

Otherwise, intended use would be for the SV.Player method to be used as a method
not a constructor:

    var player = SV.Player('my-video-id');
    // Instead of
    var player = new SV.Player('my-video-id');